+++ pj4/src/page/plant/PlantRepotting.jsx
@@ -219,6 +219,7 @@ const RepottingContent = ({
 
 // 메인 컴포넌트
 const PlantRepotting = () => {
+  const { showConfirm } = useCmDialog();
   const navigate = useNavigate();
   const location = useLocation();
   const { showAlert } = useCmDialog();
@@ -301,7 +302,8 @@ const PlantRepotting = () => {
       repottingUpdateLogs(formData)
         .unwrap()
         .then((res) => {
-          showAlert(res.message);
+          // showAlert(res.message);
+          showAlert("수정 성공!");
           setRepottingDate(dayjs());
           setSoilConditionText("");
           setRepottingMemoText("");
@@ -317,7 +319,8 @@ const PlantRepotting = () => {
       saveRepottingInfo(formData)
         .unwrap()
         .then((res) => {
-          showAlert(res.message);
+          // showAlert(res.message);
+          showAlert("저장 성공!");
           setRepottingDate(dayjs());
           setSoilConditionText("");
           setRepottingMemoText("");
@@ -333,20 +336,31 @@ const PlantRepotting = () => {
   };
 
   const handleDeleteLog = async (id) => {
-    try {
-      await deleteRepottingLogs(id).unwrap(); // 삭제 요청
-      showAlert("일지가 성공적으로 삭제되었습니다.");
+    showConfirm(
+      "알람을 삭제하시겠습니까?",
+      async () => {
+        // yes callback - 실행
+        console.log("실행 확인");
+        try {
+          await deleteRepottingLogs(id).unwrap(); // 삭제 요청
+          showAlert("삭제 성공!");
 
-      setRepottingDate(dayjs());
-      setSoilConditionText("");
-      setRepottingMemoText("");
-      setEditingLog(true);
+          setRepottingDate(dayjs());
+          setSoilConditionText("");
+          setRepottingMemoText("");
+          setEditingLog(true);
 
-      refetch();
-    } catch (error) {
-      console.error("삭제실패:", error);
-      showAlert("삭제 중 오류 발생");
-    }
+          refetch();
+        } catch (error) {
+          console.error("삭제실패:", error);
+          showAlert("삭제 중 오류 발생");
+        }
+      },
+      () => {
+        // no callback - 취소
+        console.log("실행 취소");
+      }
+    );
   };
 
   const handleEditLog = (id) => {
